test: reorganize app e2e specifications
