+++ docs/e2e/navigation.md
@@ -1,0 +1,33 @@
+# Navigation E2E テスト仕様書
+
+## 目的
+
+存在しない route を直接開いた場合でも、利用者が既定の Deck 一覧へ復帰できることを確認する。
+
+## テストケース
+
+| ID | カテゴリ | テストケース |
+| --- | --- | --- |
+| NAVIGATION-01 | read | [存在しない route から Deck 一覧へ復帰できる](#navigation-01) |
+
+<a id="navigation-01"></a>
+
+### NAVIGATION-01 存在しない route から Deck 一覧へ復帰できる
+
+カテゴリ: `read`
+
+Given:
+
+- Fixture: [`empty`](./fixture/empty.yaml)
+- 認証済みユーザーがアプリケーションにアクセスできる。
+- Sample Deck の自動生成が無効である。
+
+When:
+
+- 存在しない route を直接開き、`Go home` を選択する。
+
+Then:
+
+- Deck 一覧へ遷移する。
+- not-found 表示が残らない。
+- browser error が発生しない。
