+++ test/e2e/navigation.spec.ts
@@ -1,0 +1,14 @@
+import { expect, test } from "./fixtures";
+
+test("NAVIGATION-01 An unknown route recovers to the Deck list", async ({ fixture, page, namespace }) => {
+  await fixture.apply(page);
+
+  await page.goto(`/${namespace.id("not-a-route")}`);
+  const notFound = page.getByRole("heading", { level: 1, name: "Page not found" });
+  await expect(notFound).toBeVisible();
+  await page.getByRole("button", { name: "Go home" }).click();
+
+  await expect(page).toHaveURL(/\/$/);
+  await expect(page.getByRole("heading", { level: 1, name: "Decks" })).toBeVisible();
+  await expect(notFound).toHaveCount(0);
+});
